feat: 회원 탈퇴 시 이미지가 없으면 S3 삭제 요청 수행하지 않도록 수정 #137

+++ src/main/java/com/habitpay/habitpay/domain/member/application/MemberDeleteService.java
@@ -24,7 +24,9 @@ public class MemberDeleteService {
 
         String imageFileName = member.getImageFileName();
         log.info("[DELETE /member] imageFileName: {}", imageFileName);
-        s3FileService.deleteImage("profiles", imageFileName);
+        if (imageFileName != null) {
+            s3FileService.deleteImage("profiles", imageFileName);
+        }
 
         member.clear();
         memberRepository.save(member);
